requirements.txt
opensearch-py==2.4.2
pdf2image==1.17.0
PyJWT==2.3.0
langchain==0.1.4
sentence_transformers==2.3.1
InstructorEmbedding==1.0.1
unstructured
yandexcloud==0.255.0
langchain-community
fastapi 
uvicorn
python-dotenv